Move Vec2I to a more logical place
