+++ src/main/java/nl/tudelft/contextproject/model/level/Room.java
@@ -14,8 +14,8 @@ import com.jme3.light.Light;
 
 import nl.tudelft.contextproject.model.entities.Entity;
 import nl.tudelft.contextproject.model.level.roomIO.RoomParser;
-import nl.tudelft.contextproject.model.level.util.Vec2I;
 import nl.tudelft.contextproject.util.Size;
+import nl.tudelft.contextproject.util.Vec2I;
 
 /**
  * Class that represent a room used in level creation.
