+++ src/main/java/nl/tudelft/contextproject/model/level/util/CorridorBeautification.java
@@ -7,6 +7,7 @@ import java.util.Random;
 
 import nl.tudelft.contextproject.model.level.MazeTile;
 import nl.tudelft.contextproject.model.level.TileType;
+import nl.tudelft.contextproject.util.Vec2I;
 
 /**
  * Corridor beautifications utility class.
