+++ src/main/java/nl/tudelft/contextproject/model/level/util/DoorLocation.java
@@ -1,5 +1,6 @@
 package nl.tudelft.contextproject.model.level.util;
 
+import nl.tudelft.contextproject.util.Vec2I;
 
 /**
  * Classes that implement this abstract class are a door in a room.
