+++ src/main/java/nl/tudelft/contextproject/model/level/util/RoomEntrancePoint.java
@@ -1,5 +1,7 @@
 package nl.tudelft.contextproject.model.level.util;
 
+import nl.tudelft.contextproject.util.Vec2I;
+
 /**
  * Entrance point of a room.
  *
