+++ src/main/java/nl/tudelft/contextproject/webinterface/NormalHandler.java
@@ -15,8 +15,8 @@ import nl.tudelft.contextproject.Main;
 import nl.tudelft.contextproject.controller.EndingController;
 import nl.tudelft.contextproject.controller.GameState;
 import nl.tudelft.contextproject.model.Game;
-import nl.tudelft.contextproject.model.level.util.Vec2I;
 import nl.tudelft.contextproject.util.QRGenerator;
+import nl.tudelft.contextproject.util.Vec2I;
 import nl.tudelft.contextproject.webinterface.util.ActionUtil;
 import nl.tudelft.contextproject.webinterface.util.EntityUtil;
 import nl.tudelft.contextproject.webinterface.util.WebUtil;
