+++ src/main/java/nl/tudelft/contextproject/webinterface/websockets/COCSocket.java
@@ -10,7 +10,7 @@ import org.eclipse.jetty.websocket.api.WriteCallback;
 import nl.tudelft.contextproject.Main;
 import nl.tudelft.contextproject.controller.GameState;
 import nl.tudelft.contextproject.model.Observer;
-import nl.tudelft.contextproject.model.level.util.Vec2I;
+import nl.tudelft.contextproject.util.Vec2I;
 import nl.tudelft.contextproject.webinterface.Action;
 import nl.tudelft.contextproject.webinterface.COCErrorCode;
 import nl.tudelft.contextproject.webinterface.WebClient;
