+++ src/test/java/nl/tudelft/contextproject/model/level/util/RoomExitPointTest.java
@@ -2,6 +2,7 @@ package nl.tudelft.contextproject.model.level.util;
 
 import static org.junit.Assert.assertEquals;
 import nl.tudelft.contextproject.model.level.Room;
+import nl.tudelft.contextproject.util.Vec2I;
 
 /**
  * Test class for Room exit points.
