+++ src/test/java/nl/tudelft/contextproject/webinterface/websockets/COCSocketTest.java
@@ -13,8 +13,8 @@ import org.junit.Test;
 
 import nl.tudelft.contextproject.Main;
 import nl.tudelft.contextproject.controller.GameState;
-import nl.tudelft.contextproject.model.level.util.Vec2I;
 import nl.tudelft.contextproject.test.TestUtil;
+import nl.tudelft.contextproject.util.Vec2I;
 import nl.tudelft.contextproject.webinterface.Action;
 import nl.tudelft.contextproject.webinterface.COCErrorCode;
 import nl.tudelft.contextproject.webinterface.NormalHandler;
